+++ src/components/QueueMessage.js
@@ -1,7 +1,21 @@
 import React from 'react'
 import PropTypes from 'prop-types'
+import classnames from 'classnames'
 
-import { Alert, Button, FormText, Input } from 'reactstrap'
+import {
+  Alert,
+  Button,
+  Card,
+  CardHeader,
+  CardBody,
+  FormText,
+  Input,
+  Nav,
+  NavItem,
+  NavLink,
+  TabContent,
+  TabPane,
+} from 'reactstrap'
 import ReactMarkdown from 'react-markdown'
 
 class QueueMessage extends React.Component {
@@ -10,12 +24,27 @@ class QueueMessage extends React.Component {
 
     this.state = {
       editing: false,
-      editedMessage: null,
+      editedMessage: '',
+      activeTab: '1',
     }
 
     this.onMessageChanged = this.onMessageChanged.bind(this)
     this.onStartEdit = this.onStartEdit.bind(this)
     this.onFinishEdit = this.onFinishEdit.bind(this)
+    this.onChangeTab = this.onChangeTab.bind(this)
+
+    this.inputRef = React.createRef()
+  }
+
+  componentDidUpdate(prevProps, prevState) {
+    // If we we just started editing or we switched to the editor tab from
+    // another tab, focus the editor input
+    const { editing, activeTab } = this.state
+    const startedEditing = !prevState.editing && editing
+    const switchedToEditor = prevState.activeTab !== '1' && activeTab === '1'
+    if ((startedEditing || switchedToEditor) && this.inputRef.current) {
+      this.inputRef.current.focus()
+    }
   }
 
   onMessageChanged(e) {
@@ -27,7 +56,7 @@ class QueueMessage extends React.Component {
   onStartEdit() {
     this.setState({
       editing: true,
-      editedMessage: this.props.message,
+      editedMessage: this.props.message || '',
     })
   }
 
@@ -40,6 +69,12 @@ class QueueMessage extends React.Component {
     })
   }
 
+  onChangeTab(tab) {
+    if (this.state.activeTab !== tab) {
+      this.setState({ activeTab: tab })
+    }
+  }
+
   render() {
     const { message, isUserCourseStaff } = this.props
     const { editing, editedMessage } = this.state
@@ -50,22 +85,81 @@ class QueueMessage extends React.Component {
       return null
     }
 
+    const handleTabClick = (e, tab) => {
+      e.stopPropagation()
+      e.preventDefault()
+      this.onChangeTab(tab)
+    }
+
+    const handleTabKeyPress = (e, tab) => {
+      if (e.key === 'Enter' || e.key === ' ') {
+        e.stopPropagation()
+        this.onChangeTab(tab)
+      }
+    }
+
     let content
     let button
     if (editing) {
       content = (
-        <>
-          <Input
-            type="textarea"
-            name="text"
-            rows="6"
-            value={editedMessage}
-            onChange={this.onMessageChanged}
-          />
-          <FormText color="muted">
-            You can use Markdown to format this message.
-          </FormText>
-        </>
+        <Card>
+          <CardHeader>
+            <Nav card tabs>
+              <NavItem>
+                <NavLink
+                  href="#"
+                  className={classnames({
+                    active: this.state.activeTab === '1',
+                  })}
+                  onClick={e => handleTabClick(e, '1')}
+                  onKeyPress={e => handleTabKeyPress(e, '1')}
+                  role="tab"
+                  aria-selected={this.state.activeTab === '1'}
+                >
+                  Edit
+                </NavLink>
+              </NavItem>
+              <NavItem>
+                <NavLink
+                  href="#"
+                  className={classnames({
+                    active: this.state.activeTab === '2',
+                  })}
+                  onClick={e => handleTabClick(e, '2')}
+                  onKeyPress={e => handleTabKeyPress(e, '2')}
+                  role="tab"
+                  aria-selected={this.state.activeTab === '2'}
+                >
+                  Preview
+                </NavLink>
+              </NavItem>
+            </Nav>
+          </CardHeader>
+          <CardBody className="p-0">
+            <TabContent activeTab={this.state.activeTab}>
+              <TabPane className="m-2" tabId="1">
+                <Input
+                  type="textarea"
+                  name="text"
+                  rows="6"
+                  value={editedMessage}
+                  onChange={this.onMessageChanged}
+                  innerRef={this.inputRef}
+                />
+                <FormText color="muted">
+                  You can use Markdown to format this message.
+                </FormText>
+              </TabPane>
+              <TabPane className="m-3" tabId="2">
+                <ReactMarkdown
+                  source={
+                    editedMessage === '' ? 'Nothing to preview' : editedMessage
+                  }
+                />
+              </TabPane>
+            </TabContent>
+          </CardBody>
+        </Card>
       )
       button = (
         <Button color="primary" onClick={this.onFinishEdit}>
@@ -91,10 +185,14 @@ class QueueMessage extends React.Component {
   }
 }
 
+QueueMessage.defaultProps = {
+  message: '',
+}
+
 QueueMessage.propTypes = {
   queueId: PropTypes.number.isRequired,
   isUserCourseStaff: PropTypes.bool.isRequired,
-  message: PropTypes.string.isRequired,
+  message: PropTypes.string,
   updateQueue: PropTypes.func.isRequired,
 }